[RFC] Move props to components' types file
