+++ src/components/Button/ButtonGroup.tsx
@@ -3,47 +3,18 @@ import * as React from 'react'
 import * as _ from 'lodash'
 
 import { UIComponent, childrenExist, customPropTypes } from '../../lib'
-import { Extendable, ShorthandRenderFunction, ShorthandValue } from '../../../types/utils'
+import { Extendable } from '../../../types/utils'
 import Button from './Button'
 import { buttonGroupBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/types'
-import {
-  UIComponentProps,
-  ChildrenComponentProps,
-  ContentComponentProps,
-} from '../../lib/commonPropInterfaces'
+
 import {
   commonUIComponentPropTypes,
   childrenComponentPropTypes,
   contentComponentPropsTypes,
 } from '../../lib/commonPropTypes'
 
-export interface ButtonGroupProps
-  extends UIComponentProps<any, any>,
-    ChildrenComponentProps,
-    ContentComponentProps {
-  /**
-   * Accessibility behavior if overridden by the user.
-   * @default buttonGroupBehavior
-   */
-  accessibility?: Accessibility
-
-  /** The buttons contained inside the ButtonGroup. */
-  buttons?: ShorthandValue[]
-
-  /** The buttons inside group can appear circular. */
-  circular?: boolean
-
-  /**
-   * A custom render iterator for rendering each of the Button.Group buttons.
-   * The default component, props, and children are available for each button.
-   *
-   * @param {React.ReactType} Component - The computed component for this slot.
-   * @param {object} props - The computed props for this slot.
-   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-   */
-  renderButton?: ShorthandRenderFunction
-}
+import { ButtonGroupProps } from './ButtonGroup.types'
 
 /**
  * A button group presents multiple related actions.
